requirements.txt
bson==0.5.10
click==8.0.4
dnspython==2.2.1
Flask==2.0.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
pymongo==4.0.2
python-dateutil==2.8.2
python-dotenv==0.19.2
six==1.16.0
Werkzeug==2.0.3
